make some validation methods fluent-friendly
