+++ src/System.CommandLine/ArgumentValidation.cs
@@ -69,6 +69,95 @@ namespace System.CommandLine
             return argument;
         }
 
+        /// <summary>
+        /// Configures the argument to accept only values representing legal file names.
+        /// </summary>
+        /// <remarks>A parse error will result, for example, if file path separators are found in the parsed value.</remarks>
+        public static Argument<T> AcceptLegalFileNamesOnly<T>(this Argument<T> argument)
+        {
+            argument.Validators.Add(static result =>
+            {
+                var invalidFileNameChars = Path.GetInvalidFileNameChars();
+
+                for (var i = 0; i < result.Tokens.Count; i++)
+                {
+                    var token = result.Tokens[i];
+                    var invalidCharactersIndex = token.Value.IndexOfAny(invalidFileNameChars);
+
+                    if (invalidCharactersIndex >= 0)
+                    {
+                        result.AddError(LocalizationResources.InvalidCharactersInFileName(token.Value[invalidCharactersIndex]));
+                    }
+                }
+            });
+
+            return argument;
+        }
+
+
+        /// <summary>
+        /// Configures the argument to accept only values representing legal file paths.
+        /// </summary>
+        public static Argument<T> AcceptLegalFilePathsOnly<T>(this Argument<T> argument)
+        {
+            argument.Validators.Add(static result =>
+            {
+                var invalidPathChars = Path.GetInvalidPathChars();
+
+                for (var i = 0; i < result.Tokens.Count; i++)
+                {
+                    var token = result.Tokens[i];
+
+                    // File class no longer check invalid character
+                    // https://blogs.msdn.microsoft.com/jeremykuhne/2018/03/09/custom-directory-enumeration-in-net-core-2-1/
+                    var invalidCharactersIndex = token.Value.IndexOfAny(invalidPathChars);
+
+                    if (invalidCharactersIndex >= 0)
+                    {
+                        result.AddError(LocalizationResources.InvalidCharactersInPath(token.Value[invalidCharactersIndex]));
+                    }
+                }
+            });
+
+            return argument;
+        }
+
+
+        /// <summary>
+        /// Configures the argument to accept only the specified values, and to suggest them as command line completions.
+        /// </summary>
+        /// <param name="values">The values that are allowed for the argument.</param>
+        public static Argument<T> AcceptOnlyFromAmong<T>(
+            this Argument<T> argument, 
+            params string[] values)
+        {
+            if (values is not null && values.Length > 0)
+            {
+                argument.Validators.Clear();
+                argument.Validators.Add(UnrecognizedArgumentError);
+                argument.CompletionSources.Clear();
+                argument.CompletionSources.Add(values);
+            }
+
+            return argument;
+            
+            void UnrecognizedArgumentError(ArgumentResult argumentResult)
+            {
+                for (var i = 0; i < argumentResult.Tokens.Count; i++)
+                {
+                    var token = argumentResult.Tokens[i];
+
+                    if (token.Symbol is null || token.Symbol == argument)
+                    {
+                        if (Array.IndexOf(values, token.Value) < 0)
+                        {
+                            argumentResult.AddError(LocalizationResources.UnrecognizedArgument(token.Value, values));
+                        }
+                    }
+                }
+            }
+        }
+
         private static void FileOrDirectoryExists<T>(ArgumentResult result)
             where T : FileSystemInfo
         {
